requirement.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.0.6
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.8
IPy==1.0
lxml==4.4.2
phone==0.4.2
pycparser==2.19
pyftpdlib==1.5.6
PyMySQL==0.9.2
pypinyin==0.36.0
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.7
qqai==0.2.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
